* upstream/master: (93 commits)
  Fix non-windows fields on system/filesystem (#21758)
  disable TestReceiveEventsAndMetadata/TestSocketCleanup/TestReceiveNewEventsConcurrently in Windows (#21750)
  Use badger code from upstream repository (#21705)
  Disable writes sync in persistent cache (#21754)
  Make API address and Shard ID required in Cloud Foundry settings (#21759)
  [CI] Support skip-ci label (#21377)
  Increase recommended memory when deploying in Cloud foundry (#21755)
  typofix for dns timeout configuration option (#21069)
  chore: create CI artifacts for DEV usage (#21645)
  [Ingest Manager] Atomic installed forgotten changelog #21780
  [Ingest Manager] Agent atomic installer (#21745)
  Add missing configuration annotations (#21736)
  [Filebeat] Add check for context.DeadlineExceeded error (#21732)
  Remove kafka partition ISR from Metricbeat docs (#21709)
  Skip flaky test with oauth2 config in httpjson input (#21749)
  Fix for azure retrieve resource by ids (#21711)
  Update obs app links (#21682)
  fix: update fleet test suite name (#21738)
  Remove dot from file.extension value in Auditbeat FIM (#21644)
  Fix leaks with metadata processors (#16349)
  ...

+++ x-pack/elastic-agent/pkg/artifact/download/fs/verifier.go
@@ -51,20 +51,23 @@ func NewVerifier(config *artifact.Config, allowEmptyPgp bool, pgp []byte) (*Veri
 
 // Verify checks downloaded package on preconfigured
 // location agains a key stored on elastic.co website.
-func (v *Verifier) Verify(programName, version string) (bool, error) {
+func (v *Verifier) Verify(programName, version, artifactName string, removeOnFailure bool) (isMatch bool, err error) {
 	filename, err := artifact.GetArtifactName(programName, version, v.config.OS(), v.config.Arch())
 	if err != nil {
 		return false, errors.New(err, "retrieving package name")
 	}
 
 	fullPath := filepath.Join(v.config.TargetDirectory, filename)
+	defer func() {
+		if removeOnFailure && (!isMatch || err != nil) {
+			// remove bits so they can be redownloaded
+			os.Remove(fullPath)
+			os.Remove(fullPath + ".sha512")
+			os.Remove(fullPath + ".asc")
+		}
+	}()
 
-	isMatch, err := v.verifyHash(filename, fullPath)
-	if !isMatch || err != nil {
-		// remove bits so they can be redownloaded
-		os.Remove(fullPath)
-		os.Remove(fullPath + ".sha512")
-		os.Remove(fullPath + ".asc")
+	if isMatch, err := v.verifyHash(filename, fullPath); !isMatch || err != nil {
 		return isMatch, err
 	}
